Modularize crate
...and make minor improvements to docs
