+++ src/proofs/log_equality.rs
@@ -140,21 +140,6 @@ impl<G: Group> LogEqualityProof<G> {
         }
     }
 
-    /// Attempts to parse the proof from `bytes`. Returns `None` if `bytes` do not represent
-    /// a well-formed proof.
-    pub fn from_bytes(bytes: &[u8]) -> Option<Self> {
-        if bytes.len() != 2 * G::SCALAR_SIZE {
-            return None;
-        }
-
-        let challenge = G::deserialize_scalar(&bytes[..G::SCALAR_SIZE])?;
-        let response = G::deserialize_scalar(&bytes[G::SCALAR_SIZE..])?;
-        Some(Self {
-            challenge,
-            response,
-        })
-    }
-
     /// Verifies this proof.
     ///
     /// # Parameters
@@ -203,6 +188,21 @@ impl<G: Group> LogEqualityProof<G> {
         G::serialize_scalar(&self.challenge, &mut bytes[..G::SCALAR_SIZE]);
         G::serialize_scalar(&self.response, &mut bytes[G::SCALAR_SIZE..]);
         bytes
+    }
+
+    /// Attempts to parse the proof from `bytes`. Returns `None` if `bytes` do not represent
+    /// a well-formed proof.
+    pub fn from_bytes(bytes: &[u8]) -> Option<Self> {
+        if bytes.len() != 2 * G::SCALAR_SIZE {
+            return None;
+        }
+
+        let challenge = G::deserialize_scalar(&bytes[..G::SCALAR_SIZE])?;
+        let response = G::deserialize_scalar(&bytes[G::SCALAR_SIZE..])?;
+        Some(Self {
+            challenge,
+            response,
+        })
     }
 }
